Moved bokeh-related funcions to visualization directory
