+++ lstchain/datachecks/__init__.py
@@ -4,19 +4,14 @@ from .dl1_checker import (
     plot_datacheck,
     plot_trigger_types,
     plot_mean_and_stddev,
-    merge_dl1datacheck_files,
-    plot_mean_and_stddev_bokeh
+    merge_dl1datacheck_files
 )
 from .containers import (
     DL1DataCheckContainer,
     count_trig_types,
     DL1DataCheckHistogramBins
 )
-from .bokehcamdisplay import (
-    CameraDisplay,
-    show_camera,
-    get_pixel_location
-)
+
